+++ commands/pause.js
@@ -1,18 +1,17 @@
 exports.run = (client, message) => {
-  const voiceChannel = message.member.voiceChannel;
+  const voiceChannel = message.member.voiceChannel ? message.member.voiceChannel : (message.guild.voiceConnection ? message.guild.voiceConnnection.channel : null);
   if (!voiceChannel || (!message.member.voiceChannel && message.author.permLevel < 2)) {
     return message.reply('You must join a voice channel to pause the stream.');
   }
-  if (!client.playlists.get(message.guild.id).dispatcher.paused()) return message.reply('Audio playback has not been paused.');
+  if (client.playlists.get(message.guild.id).dispatcher.paused) return message.reply('Playback has already been paused');
   message.channel.send('Pausing playback.');
   client.playlists.get(message.guild.id).dispatcher.pause();
 };
 
 exports.conf = {
   enabled: true,
-  guildOnly: false,
   aliases: ['p'],
-  permLevel: 2
+  permLevel: 0
 };
 
 exports.help = {