+++ commands/reboot.js
@@ -34,7 +34,7 @@ exports.conf = {
   enabled: true,
   guildOnly: false,
   aliases: [],
-  permLevel: 0
+  permLevel: 4
 };
 
 exports.help = {