+++ functions/playNext.js
@@ -13,7 +13,6 @@ const playNext = (message) => {
     const embed = new Discord.RichEmbed()
       .setTitle(`Now playing **${nextSong.songTitle}** (${nextSong.playTime})`)
       .setColor(0xDD2825)
-      //.setFooter(`Requested by ${nextSong.requester}`, nextSong.requesterIcon)
       .setImage(`https://i.ytimg.com/vi/${nextSong.id}/mqdefault.jpg`)
       .setTimestamp()
       .setURL(nextSong.url);